requirements.txt
streamlit
pillow
numpy
moviepy
